+++ instrumentation/servlet/servlet-3.0-no-wrapping/src/test/java/io/opentelemetry/javaagent/instrumentation/hypertrace/servlet/v3_0/nowrapping/TestServlet.java
@@ -1,0 +1,59 @@
+/*
+ * Copyright The Hypertrace Authors
+ *
+ * Licensed under the Apache License, Version 2.0 (the "License");
+ * you may not use this file except in compliance with the License.
+ * You may obtain a copy of the License at
+ *
+ *     http://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS IS" BASIS,
+ * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ * See the License for the specific language governing permissions and
+ * limitations under the License.
+ */
+
+package io.opentelemetry.javaagent.instrumentation.hypertrace.servlet.v3_0.nowrapping;
+
+import java.io.IOException;
+import javax.servlet.http.HttpServlet;
+import javax.servlet.http.HttpServletRequest;
+import javax.servlet.http.HttpServletResponse;
+
+public class TestServlet extends HttpServlet {
+
+  public static final String RESPONSE_BODY = "{\"key\": \"val\"}";
+  public static final String RESPONSE_HEADER = "responseheader";
+  public static final String RESPONSE_HEADER_VALUE = "responsevalue";
+
+  @Override
+  protected void service(HttpServletRequest req, HttpServletResponse resp) {
+    resp.setStatus(200);
+
+    try {
+      System.out.println("Reader name");
+      System.out.println(req.getReader().getClass().getName());
+      while (req.getReader().readLine() != null) {}
+      req.getReader().close();
+    } catch (IOException e) {
+      e.printStackTrace();
+    }
+
+    resp.setContentType("application/json");
+    resp.setHeader(RESPONSE_HEADER, RESPONSE_HEADER_VALUE);
+    try {
+      //      for (int i = 0; i < RESPONSE_BODY.length(); i++) {
+      //        System.out.println("Writing a char to printWriter");
+      //        resp.getWriter().write(RESPONSE_BODY.charAt(i));
+      //      }
+      //      resp.getWriter().write(RESPONSE_BODY.toCharArray(), 0, RESPONSE_BODY.length());
+      resp.getWriter().print(RESPONSE_BODY);
+      //      resp.getWriter().print(RESPONSE_BODY);
+      //            resp.getOutputStream().print(RESPONSE_BODY);
+      //      resp.getOutputStream().close();
+    } catch (IOException e) {
+      e.printStackTrace();
+    }
+  }
+}
