+++ instrumentation/servlet/servlet-3.0-no-wrapping/src/test/resources/META-INF/services/io.opentelemetry.javaagent.tooling.InstrumentationModule
@@ -1,0 +1,1 @@
+io.opentelemetry.javaagent.instrumentation.hypertrace.servlet.v3_0.nowrapping.server.ServletInputStreamContextAccessInstrumentationModule
