--- instrumentation/servlet/servlet-3.1-no-wrapping/src/main/java/io/opentelemetry/javaagent/instrumentation/hypertrace/servlet/v3_1/nowrapping/Servlet31InstrumentationName.java
@@ -1,24 +1,0 @@
-/*
- * Copyright The Hypertrace Authors
- *
- * Licensed under the Apache License, Version 2.0 (the "License");
- * you may not use this file except in compliance with the License.
- * You may obtain a copy of the License at
- *
- *     http://www.apache.org/licenses/LICENSE-2.0
- *
- * Unless required by applicable law or agreed to in writing, software
- * distributed under the License is distributed on an "AS IS" BASIS,
- * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
- * See the License for the specific language governing permissions and
- * limitations under the License.
- */
-
-package io.opentelemetry.javaagent.instrumentation.hypertrace.servlet.v3_1.nowrapping;
-
-public class Servlet31InstrumentationName {
-  public static final String PRIMARY = "servlet-no-wrapping";
-  public static final String[] OTHER = {
-    "servlet-3-no-wrapping", "ht", "servlet-no-wrapping-ht", "servlet-no-wrapping-3-ht"
-  };
-}
