--- instrumentation/servlet/servlet-3.1-no-wrapping/src/test/resources/META-INF/services/io.opentelemetry.javaagent.tooling.InstrumentationModule
@@ -1,1 +1,0 @@
-io.opentelemetry.javaagent.instrumentation.hypertrace.servlet.v3_1.nowrapping.server.ServletInputStreamContextAccessInstrumentationModule
